requirements.txt
anyascii==0.3.2
asgiref==3.8.1
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==4.2.18
django-extra-views==0.14.0
django-filter==24.3
django-haystack==3.3.0
django-modelcluster==6.4
django-oscar==3.2.5
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-taggit==6.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
factory-boy==3.2.1
Faker==35.2.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.2
openpyxl==3.1.5
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
pillow_heif==0.21.0
purl==1.6
pysolr==3.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sorl-thumbnail==12.9.0
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wagtail==6.3.2
Willow==1.9.0
